divide the postSlice into separete file
