+++ src/components/Home.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import { deletePost, fetchPosts, updatePost } from '../redux/postSlice';
+import { deletePost, fetchPosts, updatePost } from '../redux/postActions';
 import EditModal from './EditeModal';
 
 const Home = () => {
